# api-spec.yaml
openapi: 3.1.0
info:
  title: BookWorm API
  version: 1.0.0
  description: Full-stack e-bookstore API — Next.js Route Handlers, PostgreSQL, JWT auth, Razorpay payment.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
...
    OrderItem:
      type: object
      properties:
        id:         { type: string, format: uuid }
        quantity:   { type: integer }
        unit_price: { type: number }
        book_id:    { type: string, format: uuid }
        title:      { type: string }
        author:     { type: string }
        cover:      { type: string }
        format:     { type: string }
    Order:
      type: object
      properties:
        id:              { type: string, format: uuid }
        status:          { type: string, enum: [pending, confirmed, cancelled] }
        subtotal:        { type: number }
        tax:             { type: number }
        delivery_charge: { type: number }
        discount:        { type: number }
        total:           { type: number }
        payment_method:  { type: string }
        gift_points_used:{ type: integer }
        created_at:      { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
paths:
...
  /api/wishlist:
    get:
      summary: Get wishlist
      tags: [Wishlist]
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: Wishlist items }
        '401': { description: Unauthorised }
    post:
      summary: Add to wishlist
      tags: [Wishlist]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId]
              properties:
                bookId: { type: string, format: uuid }
      responses:
        '201': { description: Added }
        '401': { description: Unauthorised }
...
  /api/orders/{orderId}:
    get:
      summary: Get single order detail
      tags: [Orders]
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Order detail, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '404': { description: Not found }
        '403': { description: Forbidden }

  # ── Checkout ─────────────────────────────────────────────────────────────────
  /api/checkout/create-order:
    post:
      summary: Create checkout order and initiate payment
      tags: [Checkout]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                addressData:
                  type: object
                  properties:
                    firstName:   { type: string }
                    lastName:    { type: string }
                    addressLine: { type: string }
                    city:        { type: string }
                    state:       { type: string }
                    pin:         { type: string }
                    phone:       { type: string }
                giftPointsToRedeem: { type: integer, default: 0 }
                paymentMethod:      { type: string, default: card }
                guestEmail:         { type: string, format: email }
      responses:
        '200':
          description: Order created — Razorpay or mock order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:       { type: string, format: uuid }
                  razorpayOrderId: { type: string }
                  amount:        { type: number }
                  currency:      { type: string }
                  isMock:        { type: boolean }
        '400': { description: Cart is empty or validation error }
  /api/checkout/verify-payment:
    post:
      summary: Verify payment and confirm order
      tags: [Checkout]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId]
              properties:
                orderId:           { type: string, format: uuid }
                razorpayOrderId:   { type: string }
                razorpayPaymentId: { type: string }
                razorpaySignature: { type: string }
      responses:
        '200':
          description: Payment verified, order confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId: { type: string, format: uuid }
                  status:  { type: string }
        '400': { description: Invalid signature or missing params }
        '404': { description: Order not found }
...